Revert "Fix/payment controller/fix for app startup"

+++ src/main/kotlin/com/itmo/microservices/demo/catalog/model/CatalogItemDto.kt
@@ -1,12 +1,11 @@
 package com.itmo.microservices.demo.catalog.model
 
-import com.itmo.microservices.demo.catalog.logic.Product
 import java.util.*
 
 data class CatalogItemDto(
     val id: UUID,
     val title: String,
     val description: String,
-    val price: Double = 100.0,
+    val price: Int = 100,
     val amount: Int
 )
